Reorganize the library into per-token-type packages
