--- anon-origin-id-test-vectors.json
@@ -1,1 +1,0 @@
-[{"sk_sign":"1d361ad638520ec433b5707f87696c2114e62b157d5ef93ef8ae7d2d25ff1a5b4a6602af57122a68f189f89495dbb6fa","pk_sign":"02a7f43d2fbf8e250868ce1053da5596963e47e1a4d9a69e3d537906777a4afd66f988b4ed0e1704e1d858d970ca7b3e0a","sk_origin":"f0765ac25117c5fff6a9db3e74e773cff3ab33c287bf64061e5a8e6c54d37148f679e7f62463f8d60599dc38621caf81","request_blind":"b1b3c1b9992d7e4aaa54957165acd61180978ddc28c98f79bd7f151270355c108e274e9ab43a055a9bc7c2e355205152","request_key":"0241e22723a128a4024d45022cfd36a4b93bb0e10cd64dffe42c36636d67f3219b5a337ecc21d58b3d44c12a19fcdc89fb","index_key":"02792433b3a27b024078bda5f36e0fd723d232448a47efb9306035599aa59e70bd763893e595462e281e4d9b13e906086e","anon_issuer_origin_id":"228dc464ce448038a2c908262947d57fc444f411f1cb4f73a2d519077edcb13a209855e0edff79a9799cd253597b69ec"}]
