--- ecdsa-blinding-test-vectors.json
@@ -1,1 +1,0 @@
-[{"Curve":"P-384","Hash":"SHA-384","skS":"4c09427c2a273b4c680aabc86e6044620146f173ad378a1063cb6d35c56d875c112fb0b5a352a911883d2cd961e8d90b","pkS":"039dc4a4d1e8e441ac68832b3b3cb25a53117e1b21434030f92754960d627839b37cd43be99640c3bef0527ee984461646","bk":"8f6d32081b61064f4c09c7c2762a0cd0caf03253ed887da91e16650d39c5bc0787b439981b9af0026f05dd37cae8c152","pkR":"022e04e5637279ed0d8cd93eaa8456b8df4e4a3f87c7a895cf6d151d260c4b1559520d687ae99f06484ba9ada55f5e0769","message":"68656c6c6f20776f726c64","context":"","signature":"2f551ca58ace75bdf1c68ee21005d674a81e60a9bf857ed535814818d43fea1792709acd53ea7a1efe8262f550bd440659ca9abc970f8d4b8598f7a81311c4f796b3651905fa28e81eac5fba449e6aec31ebf1a84aff7542f7c8fdef93b738c1"},{"Curve":"P-384","Hash":"SHA-384","skS":"c7966e96730b3d336c7227c64c68992982ee58d0502c8ee05c1cf5d24f037865403cccad3962c65f7cc71c048f399051","pkS":"0320b80a1e592ef2c386acc2e7bdd873d47acca9c89fb1723c36d1e6a9f35dccee58d74c252fe27a74ed4dbfb8ebab49de","bk":"9145e5fcac02c3b68bfe491e3e3d88e9e631b37993bce70c2670bd624f121a91ee83195258931e2f18beb29ae750991f","pkR":"027ca135f2380a535b19c1274d40ba5b41098b00271fe5561d671fbb75a98012b06fcb6f2b27f3212383c04cccdc1977ea","message":"68656c6c6f20776f726c64","context":"a135d9e39512ba4aa0624b7858a3d4ae09353bf7614ade4a4e0bf07198ea2f9a","signature":"c869b188a4b51fad78f75a38274c3ff31de46d81f35c2ba7ef2befa6223e6f0f8727aa8b967707e2df3da98d3b10e8e8549422269fa79178905b52316e1b330d0d24dc3b9f284d269c65574d805beaf030bd8b21c47b04f9bbb3cf7948cd98d9"}]
