--- ed25519-blinding-test-vectors.json
@@ -1,1 +1,0 @@
-[{"skS":"144cfb2170b3680cb63b260a61a37050f1e6b7f4fd2f4deaf5e0bb7a25ff56bd","pkS":"fd8a80317247d4d22cf762fed113c4d6f8bd263f33daaf17d03735cabd026545","bk":"2222f72709e24c5b257a07de0b719a4d31ba203bad80c6e406acecc1dc8fbb3c","pkB":"","pkR":"527135c1f6c40db8b029cc25afe8c0de8d604228d56892cd20fdb3b813d25982","message":"68656c6c6f20776f726c64","context":"","signature":"17eea8f532d43a74b82a6362b347576367a89ee471eaf06ca58c03c0b25d485de2afc2c7b0fbabc01afdbd0bcc04f30d85c68e1fcaa5e5b2e54daca2f3698d05"},{"skS":"e484f9dea13fd345c82048274b59db93a48ac3e05cbc11f318c7c41d4ecfcc67","pkS":"c3cce857791e529622a7dd878a767e2ed0ba5f633e58ff3437f3f95f9ca5486e","bk":"0000000000000000000000000000000000000000000000000000000000000000","pkB":"","pkR":"bc8d53c9ec8ec5d3d97ab0662b00b29682cadf9a3c8f9484f3357896a11a4f02","message":"68656c6c6f20776f726c64","context":"","signature":"a5c0aa1a084f9170438c7b32f28352332e3104a19772b6512b2170a1b3111f7cee8b7353c557e771490f6b89a626874265e7d16b21825d4bf66d2a881233ee05"},{"skS":"db8ad5ce54926280fc9953a745489e430d5938b73e6df030808b0048f00010c1","pkS":"a8fb6d634c20613f6fa17c3608091b6092a1ad446a7d999d9399f14193e12536","bk":"830220af05e873975248d704f8189979800177664da71ce04544eb216411c0bf","pkB":"","pkR":"db1a46554429106e3917211ce0390fbdb0d35aa2ac905d9bd9ce1de81c53f95c","message":"68656c6c6f20776f726c64","context":"620af56c20a817cc1158f6a3a02d69cc535b06ffefecbed1ae791bf938d6b73a","signature":"c1f616f28908bd6312ba3bd0ab89affabfe89b8e4fc45f1a24f824c7d84f976a9154f472bde8e448c8dbf92f214ea0cb19b6cad7fe49ed350210266890f41405"},{"skS":"69b91c842b7eeaa37e9929a083a9c327649907affe2bf1561c6117a3f1d0cc42","pkS":"4f478786eff4f1e05444c16293baa82c21569e350daa78a61c937a81de661ecd","bk":"0000000000000000000000000000000000000000000000000000000000000000","pkB":"","pkR":"e1ce6a9bc7efa8d49c9cc3922375173003a3b0026b1ab67b1f07c7391f4b925a","message":"68656c6c6f20776f726c64","context":"5e8ce68fdb6671c493213d1b4021b419c7b278026b61b277c70f220eac59c7a5","signature":"410f614549e20b25216cbd0e4b29ef0df9fc953835d8f88f7a7b4c7d20e3b21148fb36f4e0703c1cb70023cb626f275643a72c85c98cdb94ffd61a0aa1b1870c"}]
