--- origin-encryption-test-vectors.json
@@ -1,1 +1,0 @@
-[{"kem_id":32,"kdf_id":1,"aead_id":1,"issuer_encap_key_seed":"cb2ca128f89f14937519754093031387a12d864bb2ec8e78a3581abcd4fda4df","issuer_encap_key":"01002099b3c79d7629dee059bba574faeb331978f0c7fd9ee0cc2215088052804c981900010001","token_type":3,"issuer_encap_key_id":"8558d0839071e0d1a2cac30371c38712eeba25bb737230d1cde1da2a5ce38d7d","request_key":"2dcd6666488efd514c0eaf3150e4bc1fef10ff93ca07f746eed29814358174cd54a2c83813b486fd774af80135d6ea29dc","token_key_id":207,"blinded_msg":"d81efbeba072d6bfe69d57153879aac20c79e3e2de07d6eae3ad93cd0ad0c5cc08947550293f36d8b2090401dd9297987f440bfd67388b9109586436b92917f07492ebfd9096fb0b7a1aca22e1b61f0e136f65536aa693e995a97ba7712887b1a5ab7aa5b7eef7b440120996511d7760dc9efbce355db5a2d7ea508af3b0a9c4cb690c183fd5838285933ffb1af4cd05bca301336cc6c4af1576c083b1b90acd90b544b907f53d63205d74683b85b57ad4a7b43c5503439027a48eb000305638bb6142e804f094b36efe3982ab1b2b835801cb0f08c1329cd9fec00a6a950d93c5bab4fa187d3d944f144fe08ae462708d698b2932f3785bb2f00d3d441315c8","origin_name":"746573742e6578616d706c65","encap_secret":"986f192944149d17a32a85de3c65eda0","encrypted_token_request":"5b76ac0d683d5332ce77e59bb002b15f62d40b44f9e16e50409e4d848cefee5c871433ee6be4c5bf684d23c5bb884527dad436b565e370b9f6dbbce350748d26d76cba716e92d05b63056a31db61aafc662e39b4d0f348b90f2468ad66d7605370468fe789f9675c480ebcaf6e7e49cbfa227bcaa27e30984d1423a7f29093c017f229272f6fa102d56ba1149afbf7fdc0986f64a8296ac442a8b200d4f4f2d72c70800c63b78d982e773437e620c8998aed95deb8c47cf568d51c20fe6632c93aebf1c39e0da899a3d55c2acf3ebe88f191acdbb2e7392774431b4069623c327c1c0f48eaac413ab162ff811eb2538f2f72abe00aac0bd2c276b9675d36cfb1e70c4d5468a7f3e563b16908c4b9a4d00ef7916a978df99b0cb0e8241bbd454bdc0aab2751c9df0442789627da41084753b2d91697fc2829a79cbabd432f762e2f267c7680f2200bbea51202163c5c006c24a2"}]
