+++ scripts/format_test_vectors.py
