--- token-test-vectors.json
@@ -1,1 +1,0 @@
-[{"token_type":2,"issuer_name":"6973737565722e6578616d706c65","redemption_context":"be603a30db059a5b42e9ed74831438e0740c0a18fc45967d59ab266d4a0a2dc1","origin_info":"6f726967696e2e6578616d706c65","nonce":"57825763310f86c26fdcda71bc887bae46a339c6c4ff7f525184e766be778f13","token_key_id":"ca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708","token_authenticator_input":"000257825763310f86c26fdcda71bc887bae46a339c6c4ff7f525184e766be778f1325139ca69bce6a331e5e926ee147c15e9af9900678fa1fbebde57146d067abb2ca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708"},{"token_type":2,"issuer_name":"6973737565722e6578616d706c65","redemption_context":"","origin_info":"6f726967696e2e6578616d706c65","nonce":"57825763310f86c26fdcda71bc887bae46a339c6c4ff7f525184e766be778f13","token_key_id":"ca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708","token_authenticator_input":"000257825763310f86c26fdcda71bc887bae46a339c6c4ff7f525184e766be778f1311e15c91a7c2ad02abd66645802373db1d823bea80f08d452541fb2b62b5898bca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708"},{"token_type":2,"issuer_name":"6973737565722e6578616d706c65","redemption_context":"","origin_info":"","nonce":"57825763310f86c26fdcda71bc887bae46a339c6c4ff7f525184e766be778f13","token_key_id":"ca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708","token_authenticator_input":"000257825763310f86c26fdcda71bc887bae46a339c6c4ff7f525184e766be778f13b741ec1b6fd05f1e95f8982906aec1612896d9ca97d53eef94ad3c9fe023f7a4ca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708"},{"token_type":2,"issuer_name":"6973737565722e6578616d706c65","redemption_context":"be603a30db059a5b42e9ed74831438e0740c0a18fc45967d59ab266d4a0a2dc1","origin_info":"","nonce":"57825763310f86c26fdcda71bc887bae46a339c6c4ff7f525184e766be778f13","token_key_id":"ca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708","token_authenticator_input":"000257825763310f86c26fdcda71bc887bae46a339c6c4ff7f525184e766be778f131ccd616a7f71ae36963758768a4606c3eb3514371c5e9c6e6a3408b7e27da659ca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708"},{"token_type":2,"issuer_name":"6973737565722e6578616d706c65","redemption_context":"be603a30db059a5b42e9ed74831438e0740c0a18fc45967d59ab266d4a0a2dc1","origin_info":"6f726967696e2e6578616d706c652c6f726967696e322e6578616d706c65","nonce":"57825763310f86c26fdcda71bc887bae46a339c6c4ff7f525184e766be778f13","token_key_id":"ca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708","token_authenticator_input":"000257825763310f86c26fdcda71bc887bae46a339c6c4ff7f525184e766be778f13efc07116cad92c27235d5afa6e6f9a5a850539678bceb3ff1b6255fb08af3d14ca572f8982a9ca248a3056186322d93ca147266121ddeb5632c07f1f71cd2708"}]
