+++ tokens/token_request.go
@@ -1,0 +1,6 @@
+package tokens
+
+type TokenRequest interface {
+	Marshal() []byte
+	Unmarshal(data []byte) bool
+}
