+++ tokens/type1/token.go
@@ -1,0 +1,23 @@
+package type1
+
+import (
+	"fmt"
+
+	"github.com/cloudflare/pat-go/tokens"
+	"golang.org/x/crypto/cryptobyte"
+)
+
+func UnmarshalPrivateToken(data []byte) (tokens.Token, error) {
+	s := cryptobyte.String(data)
+
+	token := tokens.Token{}
+	if !s.ReadUint16(&token.TokenType) ||
+		!s.ReadBytes(&token.Nonce, 32) ||
+		!s.ReadBytes(&token.Context, 32) ||
+		!s.ReadBytes(&token.KeyID, 32) ||
+		!s.ReadBytes(&token.Authenticator, 48) {
+		return tokens.Token{}, fmt.Errorf("invalid Token encoding")
+	}
+
+	return token, nil
+}
