+++ tokens/type1/token_request.go
@@ -1,0 +1,57 @@
+package type1
+
+import (
+	"bytes"
+
+	"golang.org/x/crypto/cryptobyte"
+)
+
+var (
+	BasicPrivateTokenType = uint16(0x0001)
+)
+
+type BasicPrivateTokenRequest struct {
+	raw        []byte
+	TokenKeyID uint8
+	BlindedReq []byte // 48 bytes
+}
+
+func (r BasicPrivateTokenRequest) Type() uint16 {
+	return BasicPrivateTokenType
+}
+
+func (r BasicPrivateTokenRequest) Equal(r2 BasicPrivateTokenRequest) bool {
+	if r.TokenKeyID == r2.TokenKeyID &&
+		bytes.Equal(r.BlindedReq, r2.BlindedReq) {
+		return true
+	}
+	return false
+}
+
+func (r *BasicPrivateTokenRequest) Marshal() []byte {
+	if r.raw != nil {
+		return r.raw
+	}
+
+	b := cryptobyte.NewBuilder(nil)
+	b.AddUint16(BasicPrivateTokenType)
+	b.AddUint8(r.TokenKeyID)
+	b.AddBytes(r.BlindedReq)
+
+	r.raw = b.BytesOrPanic()
+	return r.raw
+}
+
+func (r *BasicPrivateTokenRequest) Unmarshal(data []byte) bool {
+	s := cryptobyte.String(data)
+
+	var tokenType uint16
+	if !s.ReadUint16(&tokenType) ||
+		tokenType != BasicPrivateTokenType ||
+		!s.ReadUint8(&r.TokenKeyID) ||
+		!s.ReadBytes(&r.BlindedReq, 48) {
+		return false
+	}
+
+	return true
+}
