+++ tokens/type2/basic_public_issuance_test.go
@@ -1,4 +1,4 @@
-package pat
+package type2
 
 import (
 	"bytes"
@@ -7,7 +7,9 @@ import (
 	"crypto/rsa"
 	"crypto/sha256"
 	"crypto/sha512"
+	"crypto/x509"
 	"encoding/json"
+	"encoding/pem"
 	"fmt"
 	"io/ioutil"
 	"os"
@@ -15,12 +17,84 @@ import (
 
 	"golang.org/x/crypto/cryptobyte"
 	"golang.org/x/crypto/hkdf"
+
+	"github.com/cloudflare/pat-go/tokens"
+	"github.com/cloudflare/pat-go/util"
 )
+
+// 2048-bit RSA private key
+const testTokenPrivateKey = `
+-----BEGIN RSA PRIVATE KEY-----
+MIIEowIBAAKCAQEAyxrta2qV9bHOATpM/KsluUsuZKIwNOQlCn6rQ8DfOowSmTrx
+KxEZCNS0cb7DHUtsmtnN2pBhKi7pA1I+beWiJNawLwnlw3TQz+Adj1KcUAp4ovZ5
+CPpoK1orQwyB6vGvcte155T8mKMTknaHl1fORTtSbvm/bOuZl5uEI7kPRGGiKvN6
+qwz1cz91l6vkTTHHMttooYHGy75gfYwOUuBlX9mZbcWE7KC+h6+814ozfRex26no
+KLvYHikTFxROf/ifVWGXCbCWy7nqR0zq0mTCBz/kl0DAHwDhCRBgZpg9IeX4Pwhu
+LoI8h5zUPO9wDSo1Kpur1hLQPK0C2xNLfiJaXwIDAQABAoIBAC8wm3c4tYz3efDJ
+Ffgi38n0kNvq3x5636xXj/1XA8a7otqdWklyWIm3uhEvjG/zBVHZRz4AC8NcUOFn
+q3+nOgwrIZZcS1klfBrAbL3PKOhj9nGOqMKQQ8HG2oRilJD9BJG/UtFyyVnBkhuW
+lJxyV0e4p8eHGZX6C56xEHuoVMbDKm9HR8XRwwTHRn1VsICqIzo6Uv/fJhFMu1Qf
++mtpa3oJb43P9pygirWO+w+3U6pRhccwAWlrvOjAmeP0Ndy7/gXn26rSPbKmWcI6
+3VIUB/FQsa8tkFTEFkIp1oQLejKk+EgUk66JWc8K6o3vDDyfdbmjTHVxi3ByyNur
+F87+ykkCgYEA73MLD1FLwPWdmV/V+ZiMTEwTXRBc1W1D7iigNclp9VDAzXFI6ofs
+3v+5N8hcZIdEBd9W6utHi/dBiEogDuSjljPRCqPsQENm2itTHzmNRvvI8wV1KQbP
+eJOd0vPMl5iup8nYL+9ASfGYeX5FKlttKEm4ZIY0XUsx9pERoq4PlEsCgYEA2STJ
+68thMWv9xKuz26LMQDzImJ5OSQD0hsts9Ge01G/rh0Dv/sTzO5wtLsiyDA/ZWkzB
+8J+rO/y2xqBD9VkYKaGB/wdeJP0Z+n7sETetiKPbXPfgAi7VAe77Rmst/oEcGLUg
+tm+XnfJSInoLU5HmtIdLg0kcQLVbN5+ZMmtkPb0CgYBSbhczmbfrYGJ1p0FBIFvD
+9DiCRBzBOFE3TnMAsSqx0a/dyY7hdhN8HSqE4ouz68DmCKGiU4aYz3CW23W3ysvp
+7EKdWBr/cHSazGlcCXLyKcFer9VKX1bS2nZtZZJb6arOhjTPI5zNF8d2o5pp33lv
+chlxOaYTK8yyZfRdPXCNiwKBgQDV77oFV66dm7E9aJHerkmgbIKSYz3sDUXd3GSv
+c9Gkj9Q0wNTzZKXkMB4P/un0mlTh88gMQ7PYeUa28UWjX7E/qwFB+8dUmA1VUGFT
+IVEW06GXuhv46p0wt3zXx1dcbWX6LdJaDB4MHqevkiDAqHntmXLbmVd9pXCGn/a2
+xznO3QKBgHkPJPEiCzRugzgN9UxOT5tNQCSGMOwJUd7qP0TWgvsWHT1N07JLgC8c
+Yg0f1rCxEAQo5BVppiQFp0FA7W52DUnMEfBtiehZ6xArW7crO91gFRqKBWZ3Jjyz
+/JcS8m5UgQxC8mmb/2wLD5TDvWw+XCfjUgWmvqIi5dcJgmuTAn5X
+-----END RSA PRIVATE KEY-----`
+
+func loadPrivateKey(t *testing.T) *rsa.PrivateKey {
+	block, _ := pem.Decode([]byte(testTokenPrivateKey))
+	if block == nil || block.Type != "RSA PRIVATE KEY" {
+		t.Fatal("PEM private key decoding failed")
+	}
+
+	privateKey, err := x509.ParsePKCS1PrivateKey(block.Bytes)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	return privateKey
+}
+
+func loadPrivateKeyForBenchmark(b *testing.B) *rsa.PrivateKey {
+	block, _ := pem.Decode([]byte(testTokenPrivateKey))
+	if block == nil || block.Type != "RSA PRIVATE KEY" {
+		b.Fatal("PEM private key decoding failed")
+	}
+
+	privateKey, err := x509.ParsePKCS1PrivateKey(block.Bytes)
+	if err != nil {
+		b.Fatal(err)
+	}
+
+	return privateKey
+}
 
 const (
 	outputBasicIssuanceTestVectorEnvironmentKey = "BASIC_PUBLIC_ISSUANCE_TEST_VECTORS_OUT"
 	inputBasicIssuanceTestVectorEnvironmentKey  = "BASIC_PUBLIC_ISSUANCE_TEST_VECTORS_IN"
 )
+
+func createTokenChallenge(tokenType uint16, redemptionContext []byte, issuerName string, originInfo []string) tokens.TokenChallenge {
+	challenge := tokens.TokenChallenge{
+		TokenType:       tokenType,
+		RedemptionNonce: make([]byte, len(redemptionContext)),
+		IssuerName:      issuerName,
+		OriginInfo:      originInfo,
+	}
+	copy(challenge.RedemptionNonce, redemptionContext)
+	return challenge
+}
 
 func TestBasicPublicIssuanceRoundTrip(t *testing.T) {
 	tokenKey := loadPrivateKey(t)
@@ -119,49 +193,17 @@ func (tva *basicIssuanceTestVectorArray) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-func mustMarshalPrivateKey(key *rsa.PrivateKey) []byte {
-	encodedKey, err := marshalTokenPrivateKey(key)
-	if err != nil {
-		panic(err)
-	}
-	return encodedKey
-}
-
-func mustUnmarshalPrivateKey(data []byte) *rsa.PrivateKey {
-	privateKey, err := unmarshalTokenPrivateKey(data)
-	if err != nil {
-		panic(err)
-	}
-	return privateKey
-}
-
-func mustMarshalPublicKey(key *rsa.PublicKey) []byte {
-	encodedKey, err := MarshalTokenKeyPSSOID(key)
-	if err != nil {
-		panic(err)
-	}
-	return encodedKey
-}
-
-func mustUnmarshalPublicKey(data []byte) *rsa.PublicKey {
-	publicKey, err := UnmarshalTokenKey(data)
-	if err != nil {
-		panic(err)
-	}
-	return publicKey
-}
-
 func (etv basicIssuanceTestVector) MarshalJSON() ([]byte, error) {
 	return json.Marshal(rawBasicIssuanceTestVector{
-		PrivateKey:    mustHex(mustMarshalPrivateKey(etv.skS)),
-		PublicKey:     mustHex(mustMarshalPublicKey(&etv.skS.PublicKey)),
-		Challenge:     mustHex(etv.challenge),
-		Nonce:         mustHex(etv.nonce),
-		Blind:         mustHex(etv.blind),
-		Salt:          mustHex(etv.salt),
-		TokenRequest:  mustHex(etv.tokenRequest),
-		TokenResponse: mustHex(etv.tokenResponse),
-		Token:         mustHex(etv.token),
+		PrivateKey:    util.MustHex(util.MustMarshalPrivateKey(etv.skS)),
+		PublicKey:     util.MustHex(util.MustMarshalPublicKey(&etv.skS.PublicKey)),
+		Challenge:     util.MustHex(etv.challenge),
+		Nonce:         util.MustHex(etv.nonce),
+		Blind:         util.MustHex(etv.blind),
+		Salt:          util.MustHex(etv.salt),
+		TokenRequest:  util.MustHex(etv.tokenRequest),
+		TokenResponse: util.MustHex(etv.tokenResponse),
+		Token:         util.MustHex(etv.token),
 	})
 }
 
@@ -172,24 +214,24 @@ func (etv *basicIssuanceTestVector) UnmarshalJSON(data []byte) error {
 		return err
 	}
 
-	etv.skS = mustUnmarshalPrivateKey(mustUnhex(nil, raw.PrivateKey))
-	pkS := mustUnmarshalPublicKey(mustUnhex(nil, raw.PublicKey))
+	etv.skS = util.MustUnmarshalPrivateKey(util.MustUnhex(nil, raw.PrivateKey))
+	pkS := util.MustUnmarshalPublicKey(util.MustUnhex(nil, raw.PublicKey))
 	if !pkS.Equal(&etv.skS.PublicKey) {
 		return fmt.Errorf("Mismatched public keys")
 	}
 
-	etv.challenge = mustUnhex(nil, raw.Challenge)
-	etv.nonce = mustUnhex(nil, raw.Nonce)
-	etv.blind = mustUnhex(nil, raw.Blind)
-	etv.salt = mustUnhex(nil, raw.Salt)
-	etv.tokenRequest = mustUnhex(nil, raw.TokenRequest)
-	etv.tokenResponse = mustUnhex(nil, raw.TokenResponse)
-	etv.token = mustUnhex(nil, raw.Token)
+	etv.challenge = util.MustUnhex(nil, raw.Challenge)
+	etv.nonce = util.MustUnhex(nil, raw.Nonce)
+	etv.blind = util.MustUnhex(nil, raw.Blind)
+	etv.salt = util.MustUnhex(nil, raw.Salt)
+	etv.tokenRequest = util.MustUnhex(nil, raw.TokenRequest)
+	etv.tokenResponse = util.MustUnhex(nil, raw.TokenResponse)
+	etv.token = util.MustUnhex(nil, raw.Token)
 
 	return nil
 }
 
-func generateBasicIssuanceTestVector(t *testing.T, client *BasicPublicClient, issuer *BasicPublicIssuer, tokenChallenge TokenChallenge) basicIssuanceTestVector {
+func generateBasicIssuanceTestVector(t *testing.T, client *BasicPublicClient, issuer *BasicPublicIssuer, tokenChallenge tokens.TokenChallenge) basicIssuanceTestVector {
 	challenge := tokenChallenge.Marshal()
 
 	nonce := make([]byte, 32)
@@ -273,7 +315,7 @@ func TestVectorGenerateBasicIssuance(t *testing.T) {
 	redemptionContext := make([]byte, 32)
 	hkdf.Read(redemptionContext)
 
-	challenges := []TokenChallenge{
+	challenges := []tokens.TokenChallenge{
 		createTokenChallenge(BasicPublicTokenType, redemptionContext, "issuer.example", []string{"origin.example"}),
 		createTokenChallenge(BasicPublicTokenType, nil, "issuer.example", []string{"origin.example"}),
 		createTokenChallenge(BasicPublicTokenType, nil, "issuer.example", []string{"foo.example,bar.example"}),
