+++ tokens/type2/token_request.go
@@ -1,0 +1,57 @@
+package type2
+
+import (
+	"bytes"
+
+	"golang.org/x/crypto/cryptobyte"
+)
+
+var (
+	BasicPublicTokenType = uint16(0x0002)
+)
+
+type BasicPublicTokenRequest struct {
+	raw        []byte
+	TokenKeyID uint8
+	BlindedReq []byte // 256 bytes
+}
+
+func (r BasicPublicTokenRequest) Type() uint16 {
+	return BasicPublicTokenType
+}
+
+func (r BasicPublicTokenRequest) Equal(r2 BasicPublicTokenRequest) bool {
+	if r.TokenKeyID == r2.TokenKeyID &&
+		bytes.Equal(r.BlindedReq, r2.BlindedReq) {
+		return true
+	}
+	return false
+}
+
+func (r *BasicPublicTokenRequest) Marshal() []byte {
+	if r.raw != nil {
+		return r.raw
+	}
+
+	b := cryptobyte.NewBuilder(nil)
+	b.AddUint16(BasicPublicTokenType)
+	b.AddUint8(r.TokenKeyID)
+	b.AddBytes(r.BlindedReq)
+
+	r.raw = b.BytesOrPanic()
+	return r.raw
+}
+
+func (r *BasicPublicTokenRequest) Unmarshal(data []byte) bool {
+	s := cryptobyte.String(data)
+
+	var tokenType uint16
+	if !s.ReadUint16(&tokenType) ||
+		tokenType != BasicPublicTokenType ||
+		!s.ReadUint8(&r.TokenKeyID) ||
+		!s.ReadBytes(&r.BlindedReq, 256) {
+		return false
+	}
+
+	return true
+}
