+++ tokens/type3/rate_limited_issuance_test.go
@@ -1,4 +1,4 @@
-package pat
+package type3
 
 import (
 	"bytes"
