+++ tokens/type3/token_request_test.go
@@ -1,4 +1,4 @@
-package pat
+package type3
 
 import (
 	"crypto/elliptic"
